+++ pages/register/index.js
@@ -1,0 +1,19 @@
+import Head from 'next/head'
+import Register from '../../src/views/Register'
+import { Flex, Heading, Input, Button, useColorMode, useColorModeValue } from '@chakra-ui/react'
+
+const Index = () => {
+
+  return(
+    <>
+      <Head>
+        <title>Storic | Log In</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <Register />
+    </>
+  )
+}
+
+export default Index
